annotations: refactor annotation app registry
